Refactor, break out debug names and helper functions
